cleaning up codes and removing unnecessory files
