--- examples/gcc_example/omp-for-dynamic_gcc.c
@@ -1,10 +1,0 @@
-#include <stdio.h>
-
-int main() {
-#pragma omp parallel for schedule(dynamic)
-    for(int i = 0; i < 10; i++)
-    {
-        printf("Hello World #%d\n", i );
-    }
-    return 0;
-}
