+++ tests/openmp/atomic.cpp
