+++ tests/openmp/critical_2.cpp
