+++ tests/openmp/single_copyprivate_2.cpp
